# OpenApi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"paths":{"/tags":{"get":{"tags":["tag-controller"],"operationId":"getTags","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagDTO"}}}}}}},"post":{"tags":["tag-controller"],"operationId":"createTag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagDTO"}}}}}}},"/expeditions":{"get":{"tags":["expedition-controller"],"operationId":"getExpeditions","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"animal","in":"query","required":false,"schema":{"type":"string"}},{"name":"dateFrom","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"dateTo","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"tags","in":"query","required":false,"schema":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpeditionDTO"}}}}}}},"post":{"tags":["expedition-controller"],"operationId":"createExpedition","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpeditionCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpeditionDTO"}}}}}}},"/expeditions/{id}/targets":{"post":{"tags":["expedition-controller"],"operationId":"addNewExpeditionTargets","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnimalDTO"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/expedition-targets/{id}/sightings":{"post":{"tags":["expedition-target-controller"],"operationId":"createSighting","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SightingDTO"}}},"required":true},"responses":{"201":{"description":"Created"}}},"delete":{"tags":["expedition-target-controller"],"operationId":"deleteLastSighting","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDetailDTO"}}},"required":true},"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EyeSpottedError"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDetailDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/tags/{id}":{"get":{"tags":["tag-controller"],"operationId":"getTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagDTO"}}}}}}},"/expeditions/{id}":{"get":{"tags":["expedition-controller"],"operationId":"getExpedition","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpeditionDTO"}}}}}}},"/expeditions/{id}/expedition-targets":{"get":{"tags":["expedition-controller"],"operationId":"getExpeditionTargets","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpeditionTargetCountDTO"}}}}}}}},"/expeditions/{id}/data":{"get":{"tags":["expedition-controller"],"operationId":"getExpeditionData","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpeditionDataDTO"}}}}}}},"/animals":{"get":{"tags":["animal-controller"],"operationId":"getAnimals","parameters":[{"name":"categoryName","in":"query","required":false,"schema":{"type":"string"}},{"name":"tagName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnimalDTO"}}}}}}}},"/animals/{id}":{"get":{"tags":["animal-controller"],"operationId":"getAnimal","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AnimalDTO"}}}}}}},"/expeditions/{id}/close":{"delete":{"tags":["expedition-controller"],"operationId":"closeExpedition","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpeditionDTO"}}}}}}},"/expedition-targets/{id}/sightings/{sightingsId}":{"delete":{"tags":["expedition-target-controller"],"operationId":"deleteSighting","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sightingsId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TagDTO":{"type":"object","properties":{"name":{"type":"string"}}},"ExpeditionCreateDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"expeditionTargets":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/ExpeditionTargetDTO"}}}},"ExpeditionTargetDTO":{"type":"object","properties":{"animalId":{"type":"integer","format":"int64"}}},"ExpeditionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}}},"AnimalDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"category":{"$ref":"#/components/schemas/CategoryDTO"},"tags":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/TagDTO"}},"imageData":{"type":"array","items":{"type":"string","format":"byte"}}}},"CategoryDTO":{"type":"object","properties":{"name":{"type":"string"}}},"SightingDTO":{"type":"object","properties":{"location":{"type":"string"},"expeditionTarget":{"$ref":"#/components/schemas/ExpeditionTargetDTO"}}},"RegisterDetailDTO":{"required":["email","password","repeatPassword","username"],"type":"object","properties":{"username":{"maxLength":2147483647,"minLength":1,"type":"string"},"email":{"maxLength":2147483647,"minLength":3,"type":"string"},"password":{"maxLength":2147483647,"minLength":1,"type":"string"},"repeatPassword":{"maxLength":2147483647,"minLength":1,"type":"string"}}},"EyeSpottedError":{"type":"object","properties":{"code":{"type":"string"}}},"LoginDetailDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"ExpeditionTargetCountDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"animalId":{"type":"integer","format":"int64"},"sightingsCount":{"type":"integer","format":"int64"}}},"Animal":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"category":{"$ref":"#/components/schemas/Category"},"tags":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"imageData":{"type":"array","items":{"type":"string","format":"byte"}}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ExpeditionDataDTO":{"type":"object","properties":{"totalCount":{"type":"integer","format":"int32"},"sightingsByDays":{"type":"array","items":{"$ref":"#/components/schemas/SightingsByDay"}},"sightingsByAnimals":{"type":"array","items":{"$ref":"#/components/schemas/SightingsByAnimal"}}}},"SightingsByAnimal":{"type":"object","properties":{"animal":{"$ref":"#/components/schemas/Animal"},"count":{"type":"integer","format":"int32"}}},"SightingsByDay":{"type":"object","properties":{"date":{"type":"string","format":"date-time"},"count":{"type":"integer","format":"int32"}}},"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}}}}}